team_bot/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-team-bot"
version = "0.1.0"
description = "Microsoft Teams bot for the Meeting Analyzer system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.22.0",
    "python-dotenv>=1.2.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.0",
    "botbuilder-core>=4.17.1,<4.18",
    "botbuilder-schema>=4.17.1,<4.18",
    "botbuilder-integration-aiohttp>=4.17.1,<4.18",
    "pydantic>=2.7.0",
    "meeting-shared-models",
    "meeting-orchestrator",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-anyio>=0.0.0",
    "anyio>=4.0.0",
]

[tool.setuptools.packages.find]
include = ["team_bot*"]
